Use async epqsql functions and dispcount transactions
To increase pipelining and to hold the lock on the db connections for
less time, use epgsql's async query support. Additionally do connection
initialization asynchronously to allow the service to start a bit
faster.

Finally, for prepared statements attempt to use the new dispcount
transactions mode that runs a fun that was casted to the worker. This
avoids the caller having to wait for the checkout response and then do a
checkin which should translate to less waiting. We also trick epgsql
into sending the result back to the elli process, not the pool process
which should also improve performance.