Fix leaking transaction lists
In transaction lists that end partway in a block the _rem_ainder of the
txn list is stored in the cursor as the block and the hash of the txn to
start looking from.

The issue was that the remaining transactions in the block were just
inlined on the next request without applying the same filters as the
original request did (actor, owned, and so on).

The fix adds a number of _rem_ queries to fetch the remainder of the
transaction list.