Bucket non hotspot specific sum/stats per hotspot
In order to be able to compare stats with other _hotspots_ vs for a single hotspot we need to bucket reward sums by hotspot as well as the time bucket.

This change adds bucketing by hotspot within each time bucket for any route that includes more than one hotspot. i.e. accounts. For a single hotspot the rewards are not bucketed by hotspot but just by time.

We also add a `hotspots/rewards/sums` which returns overall stats for all hotspots for a given time_range.