Change definition of online for hotspots (#416)
* Change definition of online for hotspots

This changes the definition of what it means to be online for a hotspot from confusing peer networkng status to, hopefully, a much simpler “onchain activity” metric.

If a hotspots has any of the following activity in the last 24 hours:
,
* Added
* Rewarded
* a witness
* a challenger
* a challengee
* transited data packets

It is considered to be online.  Otherwise it is considered to be ofline.

Note the slightly questionable use of “a challengee”, since he hotspot in question is not really involved in being a challengee it may not be a good metric. That said it’s the one we have today given how ETL registers changes to the gateway table.